+++ src/lib/lisp/error.c
@@ -7,6 +7,7 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdbool.h>
+#include <setjmp.h>
 #ifndef __CC65__
 #include <signal.h>
 #endif