+++ src/lib/lisp/gc.c
@@ -5,6 +5,7 @@
 #include <ctype.h>
 #include <stdbool.h>
 #include <stdlib.h>
+#include <setjmp.h>
 #ifdef TARGET_UNIX
 #include <strings.h>
 #endif
@@ -20,46 +21,9 @@
 #include <stdio.h>
 #endif
 
-extern lispptr lisp_fnin;
-extern lispptr lisp_fnout;
-
 #ifdef COMPRESSED_CONS
 bool do_compress_cons;
 #endif
-
-lispptr * global_pointers[] = {
-    &universe, &t,
-    &delayed_eval,
-    &block_sym,
-    &quote, &quasiquote, &unquote, &unquote_spliced,
-    &return_sym, &return_name, &return_value,
-    &go_sym, &go_tag,
-    &current_expr,
-#ifndef NAIVE
-    &current_toplevel,
-#endif
-    &current_function,
-    &unexpanded_toplevel,
-    &unevaluated_arg1,
-#ifndef NO_DEBUGGER
-    &onerror_sym,
-    &debug_step,
-    &breakpoints_sym,
-    &debugger_return_value_sym,
-#endif
-#ifndef NO_MACROEXPAND
-    &macroexpand_sym,
-#endif
-    &highlighted,
-
-    &lisp_fnin, &lisp_fnout,
-
-    // To be safe:
-    &x, &args, &argdefs, &arg1, &arg2, &arg2c,
-    &list_start, &list_last,
-    &value, &va,
-    NULL
-};
 
 lispptr ** gp;