+++ src/lib/lisp/io.c
@@ -6,6 +6,7 @@
 #include <string.h>
 #include <stdbool.h>
 #include <stdlib.h>
+#include <setjmp.h>
 
 #include <simpleio/libsimpleio.h>
 #include <lisp/liblisp.h>